tools/downloader: factor out the code for running parallel jobs from the converter
And move it into common.py, for subsequent use by the downloader.
